feat: implement full-stack training and course management system with progress tracking and video playback support.

+++ backend/schemas.py
@@ -1957,7 +1957,7 @@ class EmployeeDailyReportBase(BaseModel):
     hoursWorked: float = 8.0
     status: str = "Submitted" # Submitted, Reviewed
     note: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     performedBy: Optional[str] = None
     userName: Optional[str] = None
 
@@ -1972,7 +1972,7 @@ class EmployeeDailyReportUpdate(BaseModel):
     hoursWorked: Optional[float] = None
     status: Optional[str] = None
     note: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     performedBy: Optional[str] = None
     userName: Optional[str] = None
